refactor: breakdown align-branch function
