+++ src/stores/view/subscriptions/on-documents-state-update.ts
@@ -2,7 +2,7 @@ import { LineageView } from 'src/view/view';
 import { DocumentsStoreAction } from 'src/stores/documents/documents-store-actions';
 import { saveNodeContent } from 'src/view/actions/keyboard-shortcuts/helpers/commands/commands/helpers/save-node-content';
 import { focusContainer } from 'src/stores/view/subscriptions/effects/focus-container';
-import { alignBranchDebounced } from 'src/stores/view/subscriptions/effects/align-branch/align-branch';
+import { alignBranch } from 'src/stores/view/subscriptions/effects/align-branch/align-branch';
 
 export const onDocumentsStateUpdate = (
     view: LineageView,
@@ -20,7 +20,7 @@ export const onDocumentsStateUpdate = (
             action.type === 'WORKSPACE/RESIZE')
     ) {
         focusContainer(view);
-        alignBranchDebounced(
+        alignBranch(
             view.documentStore.getValue(),
             view.viewStore.getValue(),
             view.container,
